1-Agent_Level_Configuration/Parenting_Agent/pyproject.toml
[project]
name = "parenting-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.1",
]
